Reimplement distribution metric with GHC prim ops
This commit attempts to address issue #41 of tibbe/ekg-core by
replacing the C code for the distribution metric with GHC prim ops.

The performance of this implementation is about half that of the
existing C code in a single-threaded benchmark; without masking the
performance is comparable.

This commit is based on the work of Travis Whitaker in PR #42 of
tibbe/ekg-core.