Ensure safety of distribution implementation
Use the "inspection-testing" library to ensure that operations on
distributions that take locks do not allocate memory.

If they did, threads running those functions could receive exceptions or
be descheduled by the runtime while holding the lock, which could result
in deadlock or severe performance degredation, respectively.

This removes the need to wrap the distribution operations in `mask`, and
thereby improves performance. The performance of this implementation is
comparable to that of the original C code in a single-threaded
benchmark; it also scales better with multiple cores.

To eliminate allocations from the distribution operations, the
operations have been rewritten entirely in terms of prim ops from
"GHC.Prim".

Warning: the "count" field of the distribution may now be limited to 31
bits on 32-bit platforms.