Change GroupId type alias from Int to Integer
Since we can now deregister metrics, we could *technically* have an
unbounded number of groups, so we need to use `Integer` instead of `Int`
to be *completely* safe.

This also meant replacing `IntMap` with the slower `Map`. This is not a
concern for us because the only performance-sensitive operation on the
store is sampling, which acts on the group map using `elems`, and the
performance of `elems` between `IntMap`s and `Map`s are
indistinguishable (at least from a quick criterion benchmark).